[verilator simutil] Move SVScoped class into dv/verilator/cpp
