+++ hw/dv/verilator/memutil_verilator.core
@@ -12,6 +12,7 @@ filesets:
     files:
       - cpp/verilator_memutil.cc
       - cpp/verilator_memutil.h: { is_include_file: true }
+      - cpp/sv_scoped.h: { is_include_file: true }
     file_type: cppSource
 
 targets:
