+++ hw/ip/otbn/dv/model/otbn_model.cc
@@ -13,27 +13,7 @@
 #include <svdpi.h>
 
 #include "iss_wrapper.h"
-
-// Candidate for hw/dv/verilator/cpp
-/**
- * Guard class for SV Scope
- *
- * This guard ensures that all functions in the context where it is instantiated
- * are executed in an SVScope. It will restore the previous scope at destruction
- * and thereby make sure that the previous scope will be restored for all paths
- * that exit the scope.
- */
-class SVScoped {
- private:
-  svScope prev_scope_ = 0;
-
- public:
-  SVScoped(std::string scopeName) : SVScoped(scopeName.c_str()) {}
-  SVScoped(const char *scopeName) : SVScoped(svGetScopeFromName(scopeName)) {}
-
-  SVScoped(svScope scope) { prev_scope_ = svSetScope(scope); }
-  ~SVScoped() { svSetScope(prev_scope_); }
-};
+#include "sv_scoped.h"
 
 extern "C" int simutil_verilator_get_mem(int index, const svBitVecVal *val);
 extern "C" int simutil_verilator_set_mem(int index, const svBitVecVal *val);
